Move notebook-sharing test to nightly
