+++ e2e/tests/nightly/notebook/notebook-create-r.spec.ts
@@ -1,8 +1,12 @@
-import { findOrCreateWorkspace, signInWithAccessToken } from 'utils/test-utils';
+import { findOrCreateWorkspace, openTab, signInWithAccessToken } from 'utils/test-utils';
 import WorkspaceDataPage from 'app/page/workspace-data-page';
 import { makeRandomName } from 'utils/str-utils';
-import { Language } from 'app/text-labels';
+import { Language, ResourceCard, Tabs } from 'app/text-labels';
 import expect from 'expect';
+import { Page } from 'puppeteer';
+import { logger } from 'libs/logger';
+import DataResourceCard from 'app/component/card/data-resource-card';
+import WorkspaceAnalysisPage from 'app/page/workspace-analysis-page';
 
 // 30 minutes.
 jest.setTimeout(30 * 60 * 1000);
@@ -12,11 +16,12 @@ describe('Create R kernel notebook', () => {
     await signInWithAccessToken(page);
   });
 
+  let workspaceUrl: string;
   const workspaceName = 'e2eCreateRKernelNotebookTest';
   const rNotebookName = makeRandomName('R');
 
   test('Run R code', async () => {
-    await findOrCreateWorkspace(page, { workspaceName });
+    await loadWorkspace(page, workspaceName);
 
     const dataPage = new WorkspaceDataPage(page);
     const notebook = await dataPage.createNotebook(rNotebookName, Language.R);
@@ -50,4 +55,57 @@ describe('Create R kernel notebook', () => {
     await notebook.save();
     expect(cell3Output).toMatch(/success$/);
   });
+
+  test('Duplicate rename delete notebook', async () => {
+    await loadWorkspace(page, workspaceName);
+
+    const analysisPage = new WorkspaceAnalysisPage(page);
+    await openTab(page, Tabs.Analysis, analysisPage);
+
+    // Start clone notebook.
+    const cloneNotebookName = `Duplicate of ${rNotebookName}`;
+    await analysisPage.duplicateNotebookViaTable(rNotebookName);
+
+    // Rename notebook clone.
+    const newNotebookName = makeRandomName('r-cloneNotebook');
+    const modalTextContents = await analysisPage.renameResourceFromTable(
+      cloneNotebookName,
+      newNotebookName,
+      ResourceCard.Notebook
+    );
+    expect(modalTextContents).toContain(`Enter new name for ${cloneNotebookName}.ipynb`);
+
+    // Notebook card with new name is found.
+    const dataResourceCard = new DataResourceCard(page);
+    let notebookExist = await dataResourceCard.resourceExistsInTable(newNotebookName);
+    expect(notebookExist).toBe(true);
+
+    // Notebook card with old name is not found.
+    notebookExist = await dataResourceCard.resourceExistsInTable(cloneNotebookName);
+    expect(notebookExist).toBe(false);
+
+    // Delete newly renamed notebook.
+    await analysisPage.deleteResourceFromTable(newNotebookName, ResourceCard.Notebook);
+    // Verify delete was successful.
+    notebookExist = await dataResourceCard.resourceExistsInTable(newNotebookName);
+    expect(notebookExist).toBe(false);
+
+    // Delete R notebook
+    await analysisPage.deleteResourceFromTable(rNotebookName, ResourceCard.Notebook);
+    await analysisPage.waitForLoad();
+  });
+
+  // Helper functions: Load previously saved URL instead clicks thru links to open workspace data page.
+  async function loadWorkspace(page: Page, workspaceName?: string): Promise<string> {
+    if (workspaceUrl !== undefined) {
+      await page.goto(workspaceUrl, { waitUntil: ['load', 'networkidle0'] });
+      logger.info(`Goto workspace URL: ${workspaceUrl}`);
+      await new WorkspaceDataPage(page).waitForLoad();
+      return;
+    }
+
+    workspaceName = await findOrCreateWorkspace(page, { workspaceName });
+    workspaceUrl = page.url(); // Save URL for load workspace directly without search.
+    return workspaceName;
+  }
 });
